+++ packages/snaptest/lib/src/snap.dart
@@ -257,7 +257,6 @@ VoidCallback setTestViewToFakeDevice(
   }
 
   if (device is WidgetTesterDevice) {
-    restore();
     return () {};
   }