+++ packages/stupid_simple_sheet/example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:liquid_glass_renderer/liquid_glass_renderer.dart';
 import 'package:stupid_simple_sheet/stupid_simple_sheet.dart';
 
 void main() async {
@@ -20,12 +21,26 @@ final stupidSimpleSheetRoutes = [
     builder: (context, state) => const MotorExample(),
   ),
   NamedRouteDef(
+    name: 'Glass Sheet',
+    path: 'glass-sheet',
+    builder: (context, data) => _GlassSheetContent(),
+    type: RouteType.custom(
+      customRouteBuilder: <T>(context, child, page) =>
+          StupidSimpleGlassSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
+        settings: page,
+        child: child,
+      ),
+    ),
+  ),
+  NamedRouteDef(
     name: 'Cupertino Sheet',
     path: 'cupertino-sheet',
     builder: (context, data) => _CupertinoSheetContent(),
     type: RouteType.custom(
       customRouteBuilder: <T>(context, child, page) =>
           StupidSimpleCupertinoSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
         settings: page,
         child: child,
       ),
@@ -38,6 +53,7 @@ final stupidSimpleSheetRoutes = [
     type: RouteType.custom(
       customRouteBuilder: <T>(context, child, page) =>
           StupidSimpleCupertinoSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
         settings: page,
         child: child,
       ),
@@ -50,6 +66,7 @@ final stupidSimpleSheetRoutes = [
     type: RouteType.custom(
       customRouteBuilder: <T>(context, child, page) =>
           StupidSimpleSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
         settings: page,
         motion: CupertinoMotion.smooth(),
         originateAboveBottomViewInset: true,
@@ -64,8 +81,9 @@ final stupidSimpleSheetRoutes = [
     type: RouteType.custom(
       customRouteBuilder: <T>(context, child, page) =>
           StupidSimpleCupertinoSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
         settings: page,
-        snappingConfig: SheetSnappingConfig.relative(
+        snappingConfig: SheetSnappingConfig(
           [0.5, 1.0],
           initialSnap: .5,
         ),
@@ -80,6 +98,7 @@ final stupidSimpleSheetRoutes = [
     type: RouteType.custom(
       customRouteBuilder: <T>(context, child, page) =>
           StupidSimpleCupertinoSheetRoute<T>(
+        backgroundSnapshotMode: RouteSnapshotMode.openAndForward,
         settings: page,
         draggable: false,
         child: child,
@@ -119,6 +138,10 @@ class MotorExample extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             CupertinoButton.filled(
+              child: Text('Glass Sheet'),
+              onPressed: () => context.navigateTo(NamedRoute('Glass Sheet')),
+            ),
+            CupertinoButton.filled(
               child: Text('Cupertino Sheet'),
               onPressed: () =>
                   context.navigateTo(NamedRoute('Cupertino Sheet')),
@@ -142,6 +165,94 @@ class MotorExample extends StatelessWidget {
             ),
           ],
         ),
+      ),
+    );
+  }
+}
+
+class _GlassSheetContent extends StatelessWidget {
+  const _GlassSheetContent();
+
+  @override
+  Widget build(BuildContext context) {
+    return CupertinoPageScaffold(
+      child: CustomScrollView(
+        slivers: [
+          PinnedHeaderSliver(
+            child: SafeArea(
+              bottom: false,
+              child: Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: IntrinsicHeight(
+                  child: Row(
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      GestureDetector(
+                        onTap: () => Navigator.of(context).pop(),
+                        child: _GlassSurface(
+                          inLayer: false,
+                          borderRadius: BorderRadius.circular(200),
+                          child: Padding(
+                            padding: EdgeInsetsGeometry.all(10),
+                            child: Icon(
+                              CupertinoIcons.xmark,
+                            ),
+                          ),
+                        ),
+                      ),
+                      GestureDetector(
+                        onTap: () => context.pushRoute(
+                          NamedRoute('Glass Sheet'),
+                        ),
+                        child: _GlassSurface(
+                          color: CupertinoColors.activeBlue,
+                          inLayer: false,
+                          borderRadius: BorderRadius.circular(200),
+                          child: Padding(
+                            padding: EdgeInsetsGeometry.symmetric(
+                                vertical: 10, horizontal: 16),
+                            child: Center(
+                              child: Text(
+                                'Another',
+                                style: CupertinoTheme.of(context)
+                                    .textTheme
+                                    .actionTextStyle
+                                    .copyWith(
+                                      fontWeight: FontWeight.w600,
+                                      color: CupertinoColors.white.withValues(),
+                                    ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ),
+          SliverSafeArea(
+              sliver: SliverMainAxisGroup(slivers: [
+            SliverToBoxAdapter(
+              child: CupertinoTextField(
+                placeholder: 'Type something...',
+              ),
+            ),
+            SliverList(
+              delegate: SliverChildBuilderDelegate(
+                (context, index) => CupertinoListTile(
+                  title: Text('Item #$index'),
+                ),
+                childCount: 50,
+              ),
+            ),
+            SliverToBoxAdapter(
+              child: CupertinoTextField(),
+            ),
+          ]))
+        ],
       ),
     );
   }
@@ -183,6 +294,75 @@ class _CupertinoSheetContent extends StatelessWidget {
             ),
           ]))
         ],
+      ),
+    );
+  }
+}
+
+class _GlassSurface extends StatelessWidget {
+  const _GlassSurface({
+    required this.borderRadius,
+    required this.inLayer,
+    required this.child,
+    this.color,
+  });
+
+  final BorderRadius borderRadius;
+  final bool inLayer;
+  final Color? color;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return LiquidStretch(
+      child: DecoratedBox(
+        decoration: ShapeDecoration(
+          shape: RoundedSuperellipseBorder(borderRadius: borderRadius),
+          shadows: [
+            BoxShadow(
+              color: CupertinoColors.black.withValues(alpha: 0.1),
+              blurStyle: BlurStyle.outer,
+              blurRadius: 8,
+              offset: Offset(0, 0),
+            ),
+          ],
+        ),
+        child: _buildContent(context),
+      ),
+    );
+  }
+
+  Widget _buildContent(
+    BuildContext context,
+  ) {
+    if (inLayer) {
+      return LiquidGlass(
+        shape: LiquidRoundedSuperellipse(
+          borderRadius: borderRadius.topLeft.x,
+        ),
+        child: GlassGlow(child: child),
+      );
+    }
+    return LiquidGlass.withOwnLayer(
+      fake: true,
+      settings: LiquidGlassSettings(
+        glassColor: color ??
+            CupertinoTheme.of(context).barBackgroundColor.withValues(alpha: .7),
+        thickness: 30,
+        ambientStrength: .1,
+        saturation: 4,
+        lightIntensity: .4,
+        blur: 4,
+      ),
+      shape: LiquidRoundedSuperellipse(
+        borderRadius: borderRadius.topLeft.x,
+      ),
+      child: GlassGlow(
+        child: IconTheme(
+          data: IconThemeData(
+              color: CupertinoTheme.of(context).textTheme.textStyle.color),
+          child: child,
+        ),
       ),
     );
   }